// src/Pepegov.Chat.Client.Web.Minimal/src/app/models/auth/user-info.ts
export class UserInfo {
    birthdate: string;
    email: string;
    email_verified: boolean;
    phone_number: string;
    phone_number_verified: boolean;
    name: string;
    given_name: string;
    family_name: string;
    updated_at: string;
    gender: string | null;
    nickname: string;
    middle_name: string | null;
    role: string[] | null;
}